fix: move result related tests to dedicated file
